+++ src/main/java/twistedgate/overengineered/common/blocks/tileentity/UniversalMotorTileEntity.java
@@ -32,7 +32,6 @@ import net.minecraftforge.energy.CapabilityEnergy;
 import net.minecraftforge.energy.IEnergyStorage;
 import twistedgate.overengineered.common.blocks.ticking.OECommonTickableTile;
 import twistedgate.overengineered.common.multiblock.UniversalMotorMultiblock;
-import twistedgate.overengineered.experimental.common.blocks.tileentity.KineticMultiblockPartTileEntity;
 import twistedgate.overengineered.utils.AABBUtils;
 
 /**
