+++ src/main/java/twistedgate/overengineered/common/multiblock/OETemplateMultiblock.java
@@ -15,7 +15,7 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplate.StructureBlockInfo;
 import twistedgate.overengineered.OverEngineered;
-import twistedgate.overengineered.experimental.common.blocks.tileentity.KineticMultiblockPartTileEntity;
+import twistedgate.overengineered.common.blocks.tileentity.KineticMultiblockPartTileEntity;
 
 public abstract class OETemplateMultiblock extends IETemplateMultiblock{
 	private final Supplier<? extends Block> baseState;
